Organized file structure for maintaing large codebase
