+++ src/Components/Navbar.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useLocation } from "react-router-dom";
-import Modal from "./Modal"; // Import the modal component
+import Modal from "../Helper/Modal"; // Import the modal component
 
 function Navbar() {
   const [prevScrollPos, setPrevScrollPos] = useState(window.scrollY);
