+++ src/Pages/Blogs/Blogs.js
@@ -5,7 +5,7 @@ import instagramIcon from "../../images/instagram.svg"
 import twitterIcon from "../../images/twitter.svg"
 import youtubeIcon from "../../images/youtube.svg"
 import mapsIcon from "../../images/maps.svg"
-import Navbar from "../../Navbar";
+import Navbar from "../../Components/Navbar";
 import '../Services/Services.css';
 
 const images = [
