+++ src/Pages/Photos/Photos.js
@@ -5,7 +5,7 @@ import instagramIcon from "../../images/instagram.svg";
 import twitterIcon from "../../images/twitter.svg";
 import youtubeIcon from "../../images/youtube.svg";
 import mapsIcon from "../../images/maps.svg";
-import Navbar from "../../Navbar";
+import Navbar from "../../Components/Navbar";
 import "../Services/Services.css";
 
 const images = ["../../images/sunmill-loc_batcheditor_fotor (1).avif", "../../images/heroLocation (1).avif", "../../images/sunshine-loc.webp"];
