--- src/style.css
@@ -1,20 +1,0 @@
-/*-----------------------------------*\
-  #index.css
-\*-----------------------------------*/
-
-/**
- * copyright 2022 codewithsadee
- * refactored 2024 marta-vilaseca
- */
-
-@import "./styles/global.css";
-@import "./styles/reset.css";
-@import "./styles/typography.css";
-@import "./styles/layout.css";
-@import "./styles/utilities.css";
-@import "./styles/navbar.css";
-@import "./styles/hero.css";
-@import "./styles/modal.css";
-@import "./styles/sections.css";
-@import "./styles/footer.css";
-@import "./styles/media-queries.css";
